requirements.txt
fastapi
uvicorn[standard]
supabase
python-docx
python-dotenv
httpx
pandas
openpyxl
python-multipart
mangum
